refactor(frontend): split the chat channel header into hooks and a presentational bar
The channel-scoped mute fix pushed ChannelHeaderWithCounterpart past the
react-doctor large-component threshold (score 98 -> 93). Extracted rather than
shuffled: useChannelSessionActions owns the close-session, reschedule, follow-up
and complete-appointment lifecycle, useChannelInfoDrawer owns the info drawer
with its cid-tagged mute override, and ChannelHeaderBar renders the header row.
isChannelMuted moves to chatContainerUtils beside isCounterpartOnline, since both
the new hook and ChannelPreviewWrapper need it.

The component drops from ~309 lines to ~116 and the score returns to 98.

Behaviour is unchanged: the mute override keeps its channel cid tag and is still
resolved during render, so it cannot leak across a conversation switch, and the
regression test covering that still passes. No effects were introduced to sync
state, and every v8 ignore comment moved with the branch it guards.

ChannelHeaderBar initially took seven boolean props, which simply traded the
large-component warning for boolean-prop-combinations. Its props are now grouped
descriptors with no top-level booleans, which also collapsed a duplicated
online && !isGroupChat && !sessionClosed expression into a single showPresence.
